More work trying to fix the things I've broken

+++ R/dev_package/src/include/exposureModel.hpp
@@ -15,7 +15,7 @@ class exposureModel
         virtual void summary();
         virtual Rcpp::IntegerVector getOffset();
         virtual void setOffset(Rcpp::IntegerVector offs);
-        int* offset;
+        double* offset;
         int* xDim;
         int* zDim;
         double* beta;
